feat(nx-governance): Re-home signal contracts under the Core boundary
