+++ packages/governance/src/core/index.ts
@@ -1,3 +1,4 @@
 export * from './models.js';
 export * from './exceptions.js';
 export * from './profile.js';
+export * from './signals.js';
