+++ packages/governance/src/signal-engine/types.ts
@@ -1,38 +1,9 @@
-import type { Category } from '../conformance-adapter/conformance-adapter.js';
-
-export type GovernanceSignalType = KnownGovernanceSignalType | (string & {});
-
-export type KnownGovernanceSignalType =
-  | 'structural-dependency'
-  | 'cross-domain-dependency'
-  | 'missing-domain-context'
-  | 'circular-dependency'
-  | 'conformance-violation'
-  | 'domain-boundary-violation'
-  | 'layer-boundary-violation'
-  | 'ownership-gap';
-
-export type GovernanceSignalSeverity = 'info' | 'warning' | 'error';
-
-export type GovernanceSignalCategory = Category | 'structure';
-
-export type GovernanceSignalSource =
-  | 'graph'
-  | 'conformance'
-  | 'policy'
-  | 'extension';
-
-export interface GovernanceSignal {
-  id: string;
-  type: GovernanceSignalType;
-  sourceProjectId?: string;
-  targetProjectId?: string;
-  relatedProjectIds: string[];
-  severity: GovernanceSignalSeverity;
-  category: GovernanceSignalCategory;
-  message: string;
-  metadata?: Record<string, unknown>;
-  source: GovernanceSignalSource;
-  sourcePluginId?: string;
-  createdAt: string;
-}
+export type {
+  GovernanceConformanceCategory,
+  GovernanceSignal,
+  GovernanceSignalCategory,
+  GovernanceSignalSeverity,
+  GovernanceSignalSource,
+  GovernanceSignalType,
+  KnownGovernanceSignalType,
+} from '../core/signals.js';
